Update README.md
Updates the README with a notice about the difference between this fork and the original master.

This fork is not published as a gem and relies on some gems that haven't been very diligent about updating their own versions. Thus, some of this fork's dependencies are referred to by their github address, and you'll need to refer to this fork by its github location.

The library also adds sinon and sinon-chai to make your unit testing life easier when you're using it.